// NenovCars.client-app/src/components/CreateAd/interfaces/ICarAdInput.ts
export interface ICarAdInput {
    carBrand: string;
    imageUrl: string;
    carModel: string;
    hp: string;
    fuel: string;
    adTitle: string;
    yearOfProduction: string;
    color: string;
    transmission: string;
    vehicleType: string;
    price: string;
    kilometre: string;
    description: string;
}